new: DUK-156: linux support (#233)
* new: add linux presets to cmake

* chg: use boost-type-index and normalize names after demangling

* fix: add missing includes and fix clang/gcc errors

* new: use xcb to implement PlatformXCB and WindowXCB

* chg: use signed short for mouse events

* fix: free key symbols on platform destructor

* new: add AGENTS.md and update README.md git guidelines

* new: support xcb in duk_rhi

* fix: remove hardcoded cmd from justfile

* new: support fullscreen option in window

* new: add xcb cursor support

* fix: warnings and format

* chg: remove forced static linking from duk_cli

* new: add position arguments to window creation

* fix: windows compilation issues

* new: use black background color for win32 windows

* fix: strip msvc prefixes from boost::type_index.pretty_name()

* fix: guard linux-specific dependencies in cmake

* fix: use python on windows and python3 on linux in the justfile

* chg: update project files

* fix: update github workflow preset name

* new: enable linux builds on github

* chg: use clang-format from pip; run format in a parallel job; cancel previous builds after push

* fix: add linux xcb dependencies before cmake config

* fix: use gcc 14 on linux to allow explicit this in c++ 23

+++ duk_serial/src/duk_serial/json.cpp
@@ -28,19 +28,19 @@ void JsonPrimitiveValue<unsigned>::read(const rapidjson::Value& json, uint32_t& 
     value = json.GetUint();
 }
 
-void JsonPrimitiveValue<long long>::write(rapidjson::Document& document, rapidjson::Value& json, const int64_t& value) {
+void JsonPrimitiveValue<int64_t>::write(rapidjson::Document& document, rapidjson::Value& json, const int64_t& value) {
     json.SetInt64(value);
 }
 
-void JsonPrimitiveValue<long long>::read(const rapidjson::Value& json, int64_t& value) {
+void JsonPrimitiveValue<int64_t>::read(const rapidjson::Value& json, int64_t& value) {
     value = json.GetInt64();
 }
 
-void JsonPrimitiveValue<unsigned long long>::write(rapidjson::Document& document, rapidjson::Value& json, const uint64_t& value) {
+void JsonPrimitiveValue<uint64_t>::write(rapidjson::Document& document, rapidjson::Value& json, const uint64_t& value) {
     json.SetUint64(value);
 }
 
-void JsonPrimitiveValue<unsigned long long>::read(const rapidjson::Value& json, uint64_t& value) {
+void JsonPrimitiveValue<uint64_t>::read(const rapidjson::Value& json, uint64_t& value) {
     value = json.GetUint64();
 }
